Moves Comments to its own folder
