+++ src/components/Video/Video.js
@@ -1,29 +1,26 @@
 import YouTube from 'react-youtube';
-import { useLocation } from "react-router-dom";
-import './Video.css'
-import Comments from './Comments';
+import { useLocation } from 'react-router-dom';
+import './Video.css';
+import Comments from '../Comments/Comments';
 
 export default function Video() {
-    const location = useLocation();
-    const video = location.state.video
-    const id = video.id.videoId
+  const location = useLocation();
+  const video = location.state.video;
+  const id = video.id.videoId;
 
-
-    const opts = {
-        height: '600',
-        width: '1000',
-        playerVars: {
-          // https://developers.google.com/youtube/player_parameters
-          autoplay: 0,
-        },
-    };
+  const opts = {
+    height: '600',
+    width: '1000',
+    playerVars: {
+      // https://developers.google.com/youtube/player_parameters
+      autoplay: 0,
+    },
+  };
 
   return (
-    <div className='videoDiv'>
-      <YouTube className='youtubeVideo' videoId={id} opts={opts}  />
+    <div className="videoDiv">
+      <YouTube className="youtubeVideo" videoId={id} opts={opts} />
       <Comments />
     </div>
-  ) 
-  
-                 
-};
+  );
+}
